+++ src/main/java/io/github/astrapi69/file/csv/CsvFileExtensions.java
@@ -40,7 +40,6 @@ import java.util.Set;
 import java.util.StringTokenizer;
 
 import io.github.astrapi69.file.write.StoreFileExtensions;
-import io.github.astrapi69.file.write.WriteFileExtensions;
 import io.github.astrapi69.io.StreamExtensions;
 import io.github.astrapi69.string.StringExtensions;
 
@@ -626,7 +625,7 @@ public final class CsvFileExtensions
 	public static void writeLinesToFile(final Collection<String> collection, final File output,
 		final String encoding) throws IOException
 	{
-		WriteFileExtensions.writeLinesToFile(collection, output, encoding);
+		StoreFileExtensions.toFile(output, collection, encoding);
 	}
 
 }