+++ src/test/java/io/github/astrapi69/file/write/WriteFileExtensionsTest.java
@@ -31,12 +31,9 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.nio.file.Files;
-import java.util.List;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
 
@@ -67,24 +64,6 @@ public class WriteFileExtensionsTest extends FileTestCase
 		super.setUp();
 		// Create a temporary file for testing
 		tempFile = File.createTempFile("test-file", ".txt");
-	}
-
-	@Test
-	@DisplayName("Test appending lines to a file")
-	public void testAppendLines() throws IOException
-	{
-		// Arrange
-		String[] linesToAppend = { "Line 1", "Line 2", "Line 3" };
-
-		// Act
-		WriteFileExtensions.appendLines(tempFile, linesToAppend);
-
-		// Assert
-		List<String> lines = Files.readAllLines(tempFile.toPath());
-		assertEquals(3, lines.size(), "Number of lines should be 3");
-		assertEquals("Line 1", lines.get(0));
-		assertEquals("Line 2", lines.get(1));
-		assertEquals("Line 3", lines.get(2));
 	}
 
 	/**
@@ -129,7 +108,7 @@ public class WriteFileExtensionsTest extends FileTestCase
 		try (InputStream inputStream = StreamExtensions.getInputStream(testFile, true);
 			OutputStream outputStream = StreamExtensions.getOutputStream(fileout, true))
 		{
-			WriteFileExtensions.write(inputStream, outputStream);
+			StreamExtensions.writeInputStreamToOutputStream(inputStream, outputStream);
 		}
 
 		actual = FileChecksumExtensions.getCheckSumAdler32(testFile);